feat: enhance eval command with default configuration and new evaluation structure
